ai-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
scikit-learn==1.6.1
pymysql==1.1.1
sqlalchemy==2.0.35
pandas==2.2.3
redis==5.1.0
pika==1.3.2
apscheduler==3.10.4
python-dotenv==1.0.1
scipy==1.15.2
